+++ src/classes/queue-keys.ts
@@ -24,7 +24,7 @@ export class QueueKeys {
       'events',
       'pc', // priority counter key
       'marker', // marker key
-      'de', // debounce key
+      'de', // deduplication key
     ].forEach(key => {
       keys[key] = this.toKey(name, key);
     });